+++ server/src/main/java/org/opensearch/monitor/os/OsProbe.java
@@ -693,7 +693,7 @@ public class OsProbe {
         final OsStats.Mem mem = new OsStats.Mem(getTotalPhysicalMemorySize(), getFreePhysicalMemorySize());
         final OsStats.Swap swap = new OsStats.Swap(getTotalSwapSpaceSize(), getFreeSwapSpaceSize());
         final OsStats.Cgroup cgroup = Constants.LINUX ? getCgroup() : null;
-        return new OsStats(System.currentTimeMillis(), cpu, mem, swap, cgroup);
+        return new OsStats.Builder().timestamp(System.currentTimeMillis()).cpu(cpu).mem(mem).swap(swap).cgroup(cgroup).build();
     }
 
     /**